Add: Auto SFX & Remove: Accuracy in Replay Data

+++ sonolus/items/EngineData.h
@@ -49,11 +49,11 @@ class EngineDataBucket{
 class EngineDataArchetypeCallback {
     public:
 
-    double index = 0;
+    int index = 0;
     double order = 0;
 
     EngineDataArchetypeCallback(){}
-    EngineDataArchetypeCallback(double index, double order): index(index), order(order){};
+    EngineDataArchetypeCallback(int index, double order): index(index), order(order){};
     EngineDataArchetypeCallback(Json::Value arr){
         if (arr.isNull() == true) return;
         index = arr["index"].asInt(), order = arr["order"].asInt();
