+++ sonolus/items/EnginePreviewData.h
@@ -6,11 +6,11 @@ using namespace std;
 class EnginePreviewDataArchetypeCallback {
     public:
 
-    double index = 0;
+    int index = 0;
     double order = 0;
 
     EnginePreviewDataArchetypeCallback(){}
-    EnginePreviewDataArchetypeCallback(double index, double order): index(index), order(order){};
+    EnginePreviewDataArchetypeCallback(int index, double order): index(index), order(order){};
     EnginePreviewDataArchetypeCallback(Json::Value arr){
         if (arr.isNull() == true) return;
         index = arr["index"].asInt(), order = arr["order"].asInt();
